:recycle:refactor:Refactor v2 logic
