+++ src/main/java/excluz/excluz/domain/event/event/service/EventService.java
@@ -5,7 +5,11 @@ import excluz.excluz.common.exception.BadRequestException;
 import excluz.excluz.common.exception.NotFoundException;
 import excluz.excluz.common.exception.UnauthorizedException;
 import excluz.excluz.common.exception.error.ErrorCode;
-import excluz.excluz.domain.event.event.dto.*;
+import excluz.excluz.domain.event.event.dto.request.EventRequestDto;
+import excluz.excluz.domain.event.event.dto.response.EventClosingResponseDto;
+import excluz.excluz.domain.event.event.dto.response.EventResponseWithEventItemDto;
+import excluz.excluz.domain.event.event.dto.response.EventResponseWithoutEventItemDto;
+import excluz.excluz.domain.event.event.dto.response.EventWithApplicantListResponseDto;
 import excluz.excluz.domain.event.eventApplicant.enums.ApplicantStatus;
 import excluz.excluz.domain.event.eventApplicant.repository.EventApplicantRepository;
 import excluz.excluz.domain.event.eventItem.dto.EventItemRequestDto;
@@ -15,7 +19,6 @@ import excluz.excluz.domain.store.store.repository.StoreRepository;
 import excluz.excluz.domain.event.event.enums.ParticipantCondition;
 import excluz.excluz.domain.event.event.enums.SelectionMethod;
 import excluz.excluz.domain.event.event.repository.EventRepository;
-import excluz.excluz.domain.streamer.repository.StreamerRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
